draw sprite to window

+++ OpenGLTutorial/Sprite.h
@@ -7,15 +7,15 @@ public:
 	Sprite();
 	~Sprite();
 
-	void init(int x, int y, int width, int height); // should this be in the constructor
+	void init(float x, float y, float width, float height); // should this be in the constructor
 
 	void draw();
 
 private:
-	int m_x;
-	int m_y;
-	int m_width;
-	int m_height;
+	float m_x;
+	float m_y;
+	float m_width;
+	float m_height;
 	GLuint m_vboID; // GLuint is guaranteed to be 32 bits unsigned
 };
